Move message senders to Kernel methods
