+++ src/index.ts
@@ -5,5 +5,4 @@
 export * from './config';
 export * from './contents';
 export * from './kernel';
-export * from './kernelhelpers';
 export * from './session';
